Split `PurchasesTests` step 11: extracted `PurchasesTransactionHandlingTests`
Follow up to #1609.
