+++ Tests/UnitTests/Purchasing/Purchases/PurchasesTests.swift
@@ -18,150 +18,6 @@ import XCTest
 
 class PurchasesTests: BasePurchasesTests {
 
-    func testDelegateIsCalledForRandomPurchaseSuccess() throws {
-        setupPurchases()
-
-        let customerInfo = try CustomerInfo(data: Self.emptyCustomerInfoData)
-        self.backend.postReceiptResult = .success(customerInfo)
-
-        let product = MockSK1Product(mockProductIdentifier: "product")
-        let payment = SKPayment(product: product)
-
-        let customerInfoBeforePurchase = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "non_subscriptions": [:]
-            ]])
-        let customerInfoAfterPurchase = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "non_subscriptions": [product.mockProductIdentifier: []]
-            ]])
-        self.backend.overrideCustomerInfoResult = .success(customerInfoBeforePurchase)
-        self.backend.postReceiptResult = .success(customerInfoAfterPurchase)
-
-        let transaction = MockTransaction()
-
-        transaction.mockPayment = payment
-
-        transaction.mockState = SKPaymentTransactionState.purchasing
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        transaction.mockState = SKPaymentTransactionState.purchased
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        expect(self.backend.postReceiptDataCalled).to(beTrue())
-        expect(self.purchasesDelegate.customerInfoReceivedCount).toEventually(equal(2))
-    }
-
-    func testDelegateIsOnlyCalledOnceIfCustomerInfoTheSame() throws {
-        setupPurchases()
-
-        let customerInfo1 = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:],
-                "original_application_version": "1.0"
-            ]
-        ])
-
-        let customerInfo2 = customerInfo1
-
-        let product = MockSK1Product(mockProductIdentifier: "product")
-        let payment = SKPayment(product: product)
-
-        let transaction = MockTransaction()
-
-        transaction.mockPayment = payment
-
-        transaction.mockState = SKPaymentTransactionState.purchasing
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        self.backend.postReceiptResult = .success(customerInfo1)
-        transaction.mockState = SKPaymentTransactionState.purchased
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        self.backend.postReceiptResult = .success(customerInfo2)
-        transaction.mockState = SKPaymentTransactionState.purchased
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        expect(self.backend.postReceiptDataCalled).to(beTrue())
-        expect(self.purchasesDelegate.customerInfoReceivedCount).toEventually(equal(2))
-    }
-
-    func testDelegateIsCalledTwiceIfCustomerInfoTheDifferent() throws {
-        setupPurchases()
-
-        let customerInfo1 = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:],
-                "original_application_version": "1.0"
-            ]
-        ])
-
-        let customerInfo2 = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "other_purchases": [:],
-                "original_application_version": "2.0"
-            ]
-        ])
-
-        let product = MockSK1Product(mockProductIdentifier: "product")
-        let payment = SKPayment(product: product)
-
-        let transaction = MockTransaction()
-
-        transaction.mockPayment = payment
-
-        transaction.mockState = SKPaymentTransactionState.purchasing
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        self.backend.postReceiptResult = .success(customerInfo1)
-        transaction.mockState = SKPaymentTransactionState.purchased
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        self.backend.postReceiptResult = .success(customerInfo2)
-        transaction.mockState = SKPaymentTransactionState.purchased
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        expect(self.backend.postReceiptDataCalled).to(beTrue())
-        expect(self.purchasesDelegate.customerInfoReceivedCount).toEventually(equal(3))
-    }
-
-    func testDoesntIgnorePurchasesThatDoNotHaveApplicationUserNames() {
-        setupPurchases()
-        let transaction = MockTransaction()
-
-        let payment = SKMutablePayment()
-        payment.productIdentifier = "test"
-
-        expect(payment.applicationUsername).to(beNil())
-
-        transaction.mockPayment = payment
-        transaction.mockState = SKPaymentTransactionState.purchased
-
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        expect(self.backend.postReceiptDataCalled).to(beTrue())
-    }
-
     func testDoesntSetWrapperDelegateToNilIfDelegateNil() {
         setupPurchases()
         purchases!.delegate = nil
@@ -297,56 +153,6 @@ class PurchasesTests: BasePurchasesTests {
 
         expect(self.trialOrIntroPriceEligibilityChecker.invokedCheckTrialOrIntroPriceEligibilityFromOptimalStore)
             .to(beTrue())
-    }
-
-    func testProductIsRemovedButPresentInTheQueuedTransaction() throws {
-        self.mockProductsManager.stubbedProductsCompletionResult = Set()
-        setupPurchases()
-        let product = MockSK1Product(mockProductIdentifier: "product")
-
-        let customerInfoBeforePurchase = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "non_subscriptions": [:]
-            ]])
-        let customerInfoAfterPurchase = try CustomerInfo(data: [
-            "request_date": "2019-08-16T10:30:42Z",
-            "subscriber": [
-                "first_seen": "2019-07-17T00:05:54Z",
-                "original_app_user_id": "app_user_id",
-                "subscriptions": [:],
-                "non_subscriptions": [product.mockProductIdentifier: []]
-            ]])
-        self.backend.overrideCustomerInfoResult = .success(customerInfoBeforePurchase)
-        self.backend.postReceiptResult = .success(customerInfoAfterPurchase)
-
-        let payment = SKPayment(product: product)
-
-        let transaction = MockTransaction()
-
-        transaction.mockPayment = payment
-
-        transaction.mockState = SKPaymentTransactionState.purchasing
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        transaction.mockState = SKPaymentTransactionState.purchased
-        self.storeKitWrapper.delegate?.storeKitWrapper(self.storeKitWrapper, updatedTransaction: transaction)
-
-        expect(self.backend.postReceiptDataCalled).to(beTrue())
-        expect(self.purchasesDelegate.customerInfoReceivedCount).toEventually(equal(2))
-    }
-
-    private func verifyUpdatedCaches(newAppUserID: String) {
-        let expectedCallCount = 2
-        expect(self.backend.getSubscriberCallCount).toEventually(equal(expectedCallCount))
-        expect(self.deviceCache.cachedCustomerInfo.count).toEventually(equal(expectedCallCount))
-        expect(self.deviceCache.cachedCustomerInfo[newAppUserID]).toEventuallyNot(beNil())
-        expect(self.purchasesDelegate.customerInfoReceivedCount).toEventually(equal(expectedCallCount))
-        expect(self.deviceCache.setCustomerInfoCacheTimestampToNowCount).toEventually(equal(expectedCallCount))
-        expect(self.mockOfferingsManager.invokedUpdateOfferingsCacheCount).toEventually(equal(expectedCallCount))
     }
 
 }
